Refactor out HmBrushEditorState
