+++ Hesiod/include/hesiod/gui.hpp
@@ -63,8 +63,6 @@ void flip_vertically(int width, int height, uint8_t *data);
 
 void help_marker(std::string text);
 
-void hmap_brush_editor(hesiod::vnode::ViewBrush::HmBrushEditorState& state, ImTextureID canvas_texture, float width = 0.0f);
-
 bool listbox_map_enum(std::map<std::string, int> &map,
                       int                        &selected,
                       float                       width);
